requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.23.0,<2.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
# --- Core AI & LLM ---
openai
python-dotenv

# --- RAG & PDF Processing ---
chromadb
pymupdf
sentence-transformers

# --- Voice: Speech-to-Text (Whisper) ---
openai-whisper
sounddevice
numpy
scipy
torch
keyboard

# --- Voice: Text-to-Speech (Kokoro/ONNX) ---
onnxruntime
misaki[en]